+++ client/src/inventory.rs
@@ -20,7 +20,7 @@ pub struct Props {
     pub inventory: MInventory,
     pub dict: Dictionary,
     pub change_panel_signal: Callback<KeyboardEvent>,
-    pub item_action_signal: Callback<(KeyboardEvent, u64, i32)>,
+    pub item_action_signal: Callback<(KeyboardEvent, i32, i32)>,
     pub target_indicator_signal: Callback<(Option<KeyboardEvent>, Option<i32>)>,
 }
 
@@ -172,7 +172,7 @@ impl Component for Inventory {
     }
 
     fn view(&self) -> Html {
-        let render_items = |item: &(i32, u64)| {
+        let render_items = |item: &(i32, i32)| {
             let name = self.props.dict.get_name(item.0);
             let css = self.props.dict.get_css(item.0);
             html! {
